+++ src/Phan/Config.php
@@ -307,6 +307,12 @@ class Config
         // Phan will not assume it knows specific types if the default value is `false` or `null`.
         'guess_unknown_parameter_type_using_default' => false,
 
+        // Allow adding types to vague return types such as @return object, @return ?mixed in function/method/closure union types.
+        // Normally, Phan only adds inferred returned types when there is no `@return` type or real return type signature..
+        //
+        // Disabled by default. This is more useful with `--analyze-twice`.
+        'allow_overriding_vague_return_types' => false,
+
         // When enabled, infer that the types of the properties of `$this` are equal to their default values at the start of `__construct()`.
         // This will have some false positives due to Phan not checking for setters and initializing helpers.
         // This does not affect inherited properties.