+++ src/Phan/Language/Type/FunctionLikeDeclarationType.php
@@ -743,6 +743,11 @@ abstract class FunctionLikeDeclarationType extends Type implements FunctionInter
         return $this->return_type;
     }
 
+    public function getOriginalReturnType(): UnionType
+    {
+        return $this->return_type;
+    }
+
     public function getSuppressIssueList(): array
     {
         // TODO: Inherit suppress issue list from phpdoc declaring this?